+++ src/resources/ui.rs
@@ -1,2 +1,3 @@
+pub mod cluster_info_window_state;
 pub mod generate_cubic_modal_state;
 pub mod generate_seed_modal_state;