+++ src/types.rs
@@ -1,2 +1,3 @@
 pub mod cluster_generation_method;
 pub mod cluster_generation_settings;
+pub mod sort_direction;