+++ src/types/sort_direction.rs
@@ -1,0 +1,6 @@
+#[derive(Debug, Default, Copy, Clone, PartialEq)]
+pub enum SortDirection {
+    #[default]
+    Ascending,
+    Descending,
+}